Move uninitialized timeline from tenant.rs to timeline/uninit.rs
